requirements.txt
flask
flask-mysql
redis
python-memcached
pymongo
